codecov: switch to informational mode (#76)
Use Codecov in informational mode. Default is false. If true is specified the
resulting status will pass no matter what the coverage is or what other
settings are specified. Informational mode is great to use if you want to
expose codecov information to other developers in your pull request without
necessarily gating PRs on that information.